Minor entities and existing DTOs changes, new DTOs added, business logic implementation of BookingServiceImpl

+++ src/main/java/com/academy/project/hotelsmanagementsystem/dto/HotelDTO.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -15,20 +17,16 @@ public class HotelDTO {
     private Long id;
     @NotNull(message = "{hotel.validations.hotelName}")
     private String hotelName;
-
     private String address;
-
     private Integer postCode;
     @NotNull(message = "{hotel.validations.city}")
     private String city;
-
     private String country;
     @NotNull(message = "{hotel.validations.roomsNr}")
     private Integer roomsNr;
     @NotNull(message = "{hotel.validations.phone}")
     @Pattern(regexp = "^\\+355\\s(68|69)\\d{7}$\n",message = "{hotel.validations.phone.regexp}")
     private String phone;
-
-    private Integer starRating;
+    private BigDecimal starRating;
 
 }
